test: extract conditional expression tests
Part of:
- #2806
